\name{PISA2000lite-package}
\alias{PISA2000lite-package}
\alias{PISA2000lite}
\docType{package}
\title{
Dataset from Programme for International Student Assessment 2000

}
\description{
This is an R package that aggregates data from PISA 2000.
Be warned that the zip is about 100MB.

From wikipedia: The Programme for International Student Assessment (PISA) is a worldwide study by the Organisation for Economic Co-operation and Development (OECD) in member and non-member nations of 15-year-old school pupils' scholastic performance on mathematics, science, and reading. It was first performed in 2000 and then repeated every three years. It is done with view to improving education policies and outcomes. The data has increasingly been used both to assess the impact of education quality on incomes and growth and for understanding what causes differences in achievement across nations. see [http://en.wikipedia.org/wiki/Programme_for_International_Student_Assessment]

The original dataset is avaliable here [http://pisa2000.acer.edu.au/downloads.php]

}
\details{
\tabular{ll}{
Package: \tab PISA2000lite\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Przemyslaw Biecek (OECD)

Maintainer: Przemyslaw Biecek <Przemyslaw.Biecek@gmail.com>
}
\keyword{package}
\seealso{
\code{link{math2000}}, 
\code{link{read2000}}, 
\code{link{scie2000}}, 
\code{link{school2000}}, 
\code{link{item2000}}.
}
\examples{
}
